// src/lib/api/models/updateAssignmentPolicyDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { UpdateAssignmentPolicyDtoTransactionType } from './updateAssignmentPolicyDtoTransactionType';
import type { UpdateAssignmentPolicyDtoSellingMode } from './updateAssignmentPolicyDtoSellingMode';
import type { UpdateAssignmentPolicyDtoExpireBehavior } from './updateAssignmentPolicyDtoExpireBehavior';
import type { UpdateAssignmentPolicyDtoStatus } from './updateAssignmentPolicyDtoStatus';

export interface UpdateAssignmentPolicyDto {
  /** Policy name */
  name?: string;
  propertyTypeId?: string;
  transactionType?: UpdateAssignmentPolicyDtoTransactionType;
  sellingMode?: UpdateAssignmentPolicyDtoSellingMode;
  projectId?: string;
  zoneId?: string;
  maxAssignedUsers?: number;
  durationDays?: number;
  autoExtendEnabled?: boolean;
  expireBehavior?: UpdateAssignmentPolicyDtoExpireBehavior;
  status?: UpdateAssignmentPolicyDtoStatus;
  priority?: number;
}